Crate offscreen_gl_context

Structs

DrawBuffer
This structure represents an offscreen context draw buffer. It has a framebuffer, with at least color renderbuffer (alpha or not). It may also have a depth or stencil buffer, depending on context requirements.

GLContext
This is a wrapper over a native headless GL context

GLContextAttributes
This structure represents the attributes the context must support It's almost (if not) identical to WebGLGLContextAttributes

GLContextCapabilities
This is a cross-platform struct, that every GLContext implementation should have under the field capabilities, as a public field This should allow us to know the capabilities of a given GLContext without repeating the same code over and over

GLFormats
This structure is here to allow cross-platform formatting

GLLimits

NativeGLContext

NativeGLContextHandle

OSMesaContext

OSMesaContextHandle

Enums

ColorAttachmentType

GLFeature

Traits

GLContextDispatcher

NativeGLContextMethods