Struct offscreen_gl_context::gl_context::GLContext

pub struct GLContext<Native> {
    native_context: Native,
    draw_buffer: Option<DrawBuffer>,
    attributes: GLContextAttributes,
    capabilities: GLContextCapabilities,
    formats: GLFormats,
    limits: GLLimits,
}

This is a wrapper over a native headless GL context

Fields

native_context

draw_buffer

This an abstraction over a custom framebuffer with attachments according to WebGLContextAttributes

attributes

capabilities

formats

limits

Methods

impl<Native> GLContext<Native> where Native: NativeGLContextMethods

fn create(shared_with: Option<&Native::Handle>) -> Result<GLContext<Native>, &'static str>

fn get_proc_address(addr: &str) -> *const ()

fn current_handle() -> Option<Native::Handle>

fn new(size: Size2D<i32>, attributes: GLContextAttributes, color_attachment_type: ColorAttachmentType, shared_with: Option<&Native::Handle>) -> Result<GLContext<Native>, &'static str>

fn with_default_color_attachment(size: Size2D<i32>, attributes: GLContextAttributes, shared_with: Option<&Native::Handle>) -> Result<GLContext<Native>, &'static str>

fn make_current(&self) -> Result<(), &'static str>

fn unbind(&self) -> Result<(), &'static str>

fn is_current(&self) -> bool

fn handle(&self) -> Native::Handle

fn borrow_attributes(&self) -> &GLContextAttributes

fn borrow_capabilities(&self) -> &GLContextCapabilities

fn borrow_formats(&self) -> &GLFormats

fn borrow_limits(&self) -> &GLLimits

fn borrow_draw_buffer(&self) -> Option<&DrawBuffer>

fn get_framebuffer(&self) -> GLuint

fn draw_buffer_size(&self) -> Option<Size2D<i32>>

fn resize(&mut self, size: Size2D<i32>) -> Result<(), &'static str>