Struct offscreen_gl_context::DrawBuffer

pub struct DrawBuffer { /* fields omitted */ }

This structure represents an offscreen context draw buffer. It has a framebuffer, with at least color renderbuffer (alpha or not). It may also have a depth or stencil buffer, depending on context requirements.

Methods

impl DrawBuffer

fn new<T: NativeGLContextMethods>(context: &GLContext<T>,
size: Size2D<i32>,
color_attachment_type: ColorAttachmentType)
-> Result<DrawBuffer, &'static str>

fn get_framebuffer(&self) -> GLuint

fn size(&self) -> Size2D<i32>

fn color_attachment_type(&self) -> ColorAttachmentType

fn get_bound_color_renderbuffer_id(&self) -> Option<GLuint>

fn get_bound_texture_id(&self) -> Option<GLuint>